Rename package to names
